requirements.txt
# AI PDF Q&A — Python dependencies
# Install with: pip install -r requirements.txt

# Core LangChain
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0

# OpenAI
openai>=1.0.0

# Vector store
qdrant-client>=1.9.0
langchain-qdrant>=0.1.3

# EnsembleRetriever (BM25 + Qdrant hybrid)
langchain-classic>=1.0.0

# PDF loading
pypdf>=4.0.0

# Streamlit UI
streamlit>=1.39.0

# BM25 keyword retrieval
rank-bm25>=0.2.2

# Tokenisation (required by OpenAI embeddings)
tiktoken>=0.7.0

# Environment variables
python-dotenv>=1.0.0
